Add support for delete/put users by admins

+++ src/main/java/student/laurens/novibackend/users/User.java
@@ -15,7 +15,7 @@ public class User {
     @Id
     @Column(name = "UID")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private @Getter Long uid;
+    private @Getter Integer uid;
 
     @Column(name = "USERNAME", nullable = false, unique = true)
     private @Getter @Setter String username;
